requirements.txt
streamlit==1.22.0
pillow==9.5.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tensorflow==2.12.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.7.0.72
matplotlib==3.7.1
easyocr==1.7.0 